refactor: 合并代码到基类
